Refactored explodable code so it can be overriden with implementation hooks
