+++ Assets/Game/Scripts/Explodables/Editor/ExplodableEditor.cs
@@ -2,14 +2,14 @@
 using UnityEngine;
 
 namespace GameJammers.GGJ2025.Explodables {
-    [CustomEditor(typeof(Explodable))]
+    [CustomEditor(typeof(ExplodableExample))]
     public class ExplodableEditor : Editor {
         public override void OnInspectorGUI() {
             base.OnInspectorGUI();
 
             // Add a button to the inspector when the game is playing
             if (Application.isPlaying) {
-                var explodable = (Explodable)target;
+                var explodable = (ExplodableExample)target;
 
                 if (GUILayout.Button("Explode")) {
                     explodable.Explode();
