--- Assets/Game/Scripts/Explodables/Explodable.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 8a93d20dec0e4d7f920b8e640d0e307e
-timeCreated: 1737743918
