+++ Assets/Game/Scripts/Explodables/ExplodableExample.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 8a93d20dec0e4d7f920b8e640d0e307e
+timeCreated: 1737743918
