+++ Assets/Game/Scripts/Explodables/TriggerExplodableTracker.cs
@@ -3,19 +3,19 @@ using UnityEngine;
 
 namespace GameJammers.GGJ2025.Explodables {
     public class TriggerExplodableTracker : MonoBehaviour {
-        readonly List<Explodable> _trackedObjects = new();
+        readonly List<ExplodableBase> _trackedObjects = new();
 
-        public IReadOnlyList<Explodable> TrackedObjects => _trackedObjects;
+        public IReadOnlyList<ExplodableBase> TrackedObjects => _trackedObjects;
 
         void OnTriggerEnter (Collider other) {
-            var explodable = other.gameObject.GetComponent<Explodable>();
+            var explodable = other.gameObject.GetComponent<ExplodableBase>();
             if (explodable) {
                 _trackedObjects.Add(explodable);
             }
         }
 
         void OnTriggerExit (Collider other) {
-            var explodable = other.gameObject.GetComponent<Explodable>();
+            var explodable = other.gameObject.GetComponent<ExplodableBase>();
             if (explodable) {
                 _trackedObjects.Remove(explodable);
             }
